feat: updated portal and browser url to be environment specific (#2885)
