+++ next/components/Home/components/Section/components/SectionHero/sectionHero.tsx
@@ -1,6 +1,5 @@
 import { ButtonPrimary } from "@clevercanary/data-explorer-ui/lib/components/common/Button/components/ButtonPrimary/buttonPrimary";
 import { ButtonSecondary } from "@clevercanary/data-explorer-ui/lib/components/common/Button/components/ButtonSecondary/buttonSecondary";
-import { PORTAL_URL } from "../../../../../../site-config/anvil-portal/dev/config";
 import { Section } from "../../section.styles";
 import { Carousel } from "./components/Carousel/carousel";
 import {
@@ -14,7 +13,11 @@ import {
 const GET_STARTED = "/learn";
 const LEARN_MORE = "/overview";
 
-export const SectionHero = (): JSX.Element => {
+interface SectionHeroProps {
+  portalURL: string;
+}
+
+export const SectionHero = ({ portalURL }: SectionHeroProps): JSX.Element => {
   return (
     <Section>
       <SectionLayout>
@@ -22,10 +25,10 @@ export const SectionHero = (): JSX.Element => {
           <Head>Migrate Your Genomic Research to the Cloud</Head>
           <Subhead>Secure, cost-effective genomic analysis at scale.</Subhead>
           <CTAs>
-            <ButtonPrimary href={`${PORTAL_URL}${GET_STARTED}`}>
+            <ButtonPrimary href={`${portalURL}${GET_STARTED}`}>
               Get Started
             </ButtonPrimary>
-            <ButtonSecondary href={`${PORTAL_URL}${LEARN_MORE}`}>
+            <ButtonSecondary href={`${portalURL}${LEARN_MORE}`}>
               Learn More
             </ButtonSecondary>
           </CTAs>
