+++ next/config/config.ts
@@ -1,7 +1,7 @@
 import { setConfig } from "@clevercanary/data-explorer-ui/lib/config/config";
-import { SiteConfig } from "@clevercanary/data-explorer-ui/lib/config/entities";
 import anvilPortalDev from "../site-config/anvil-portal/dev/config";
 import anvilPortalProd from "../site-config/anvil-portal/prod/config";
+import { SiteConfig } from "../site-config/common/entities";
 
 const CONFIGS: { [k: string]: SiteConfig } = {
   "anvil-portal-dev": anvilPortalDev,
