+++ next/providers/config.tsx
@@ -1,0 +1,28 @@
+import { getDefaultConfig } from "@clevercanary/data-explorer-ui/lib/config/utils";
+import { createContext, ReactNode } from "react";
+import { SiteConfig } from "../site-config/common/entities";
+
+export type ConfigContextProps = {
+  config: SiteConfig;
+};
+
+export interface ConfigProps {
+  children: ReactNode | ReactNode[];
+  config: SiteConfig;
+}
+
+export const ConfigContext = createContext<ConfigContextProps>({
+  config: { ...getDefaultConfig(), portalURL: "" },
+});
+
+export function ConfigProvider({ children, config }: ConfigProps): JSX.Element {
+  return (
+    <ConfigContext.Provider
+      value={{
+        config,
+      }}
+    >
+      {children}
+    </ConfigContext.Provider>
+  );
+}
